# backend/swagger.yaml
openapi: 3.0.3
info:
  title: TaxLaw API Documentation
  description: API system for document management, retrieval, and conversation.
  version: 1.0.0

servers:
  - url: http://127.0.0.1:5000

tags:
  - name: Auth
  - name: Documents
  - name: Admin
  - name: Conversation
  - name: Retrieval

paths:

  # ============================
  # AUTH
  # ============================
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    $ref: "#/components/schemas/User"
        401:
          description: Wrong email or password

  # ============================
  # DOCUMENTS
  # ============================

  /documents/create:
    post:
      tags: [Documents]
      security: [{ BearerAuth: [] }]
      summary: Upload new document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                tax_type: { type: string }
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/list:
    get:
      tags: [Documents]
      security: [{ BearerAuth: [] }]
      summary: List all documents
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"

  /documents/{id}:
    get:
      tags: [Documents]
      security: [{ BearerAuth: [] }]
      summary: Get document details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

    delete:
      tags: [Documents]
      security: [{ BearerAuth: [] }]
      summary: Delete a document (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: Deleted
          content:
            application/json:
              example:
                message: Document deleted

  /documents/{id}/file:
    get:
      tags: [Documents]
      summary: View or download PDF
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: Return PDF file
          content:
            application/pdf: {}

  /documents/{id}/audit:
    get:
      tags: [Documents]
      security: [{ BearerAuth: [] }]
      summary: Get audit log for a document
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: List of audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

  /documents/{id}/status:
    patch:
      tags: [Documents]
      security: [{ BearerAuth: [] }]
      summary: Update status (admin & lawyer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
                feedback: { type: string }
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: Status updated
          content:
            application/json:
              example:
                message: Status updated

  # ============================
  # ADMIN
  # ============================

  /admin/users:
    get:
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      summary: Get list of users
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  # ============================
  # CONVERSATION
  # ============================

  /conversation/send:
    post:
      tags: [Conversation]
      security: [{ BearerAuth: [] }]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversation_id: { type: integer }
                message: { type: string }
      responses:
        200:
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /conversation/{id}/messages:
    get:
      tags: [Conversation]
      security: [{ BearerAuth: [] }]
      summary: Get messages of a conversation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  # ============================
  # RETRIEVAL
  # ============================

  /retrieval/query:
    post:
      tags: [Retrieval]
      security: [{ BearerAuth: [] }]
      summary: Query retrieval system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query: { type: string }
      responses:
        200:
          description: Retrieval result
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items:
                      $ref: "#/components/schemas/RetrievalMatch"

  /retrieval/logs:
    get:
      tags: [Retrieval]
      security: [{ BearerAuth: [] }]
      summary: Get retrieval logs
      responses:
        200:
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RetrievalMatch"

# =========================================
# COMPONENT SCHEMAS
# =========================================

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string }

    Document:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        tax_type: { type: string }
        status: { type: string }
        lawyer_feedback: { type: string }
        created_at: { type: string, format: date-time }

    AuditLog:
      type: object
      properties:
        id: { type: integer }
        document_id: { type: integer }
        action: { type: string }
        old_status: { type: string }
        new_status: { type: string }
        user_id: { type: integer }
        message: { type: string }
        created_at: { type: string, format: date-time }

    Message:
      type: object
      properties:
        id: { type: integer }
        conversation_id: { type: integer }
        message: { type: string }
        created_at: { type: string }

    RetrievalMatch:
      type: object
      properties:
        id: { type: integer }
        query: { type: string }
        score: { type: number }
        document_id: { type: integer }